check if dir exists before clearing

+++ build.ts
@@ -3,16 +3,20 @@ import nodeExternalsPlugin from "esbuild-node-externals"
 import klawSync from "klaw-sync"
 import rimraf from "rimraf"
 import { exec } from "child_process"
+import { existsSync } from "fs"
 const args = process.argv.slice(2)
 const watch = args.includes("watch")
 const root = process.cwd() + "/lib"
 
-const lib = klawSync(root)
-lib.forEach(({ path }) => {
-    if (!path.endsWith("package.json")) {
-        rimraf.sync(path)
-    }
-})
+if (existsSync(root)) {
+    console.log("CLEARING LIB DIR")
+    const lib = klawSync(root)
+    lib.forEach(({ path }) => {
+        if (!path.endsWith("package.json")) {
+            rimraf.sync(path)
+        }
+    })
+}
 console.log("RUNNING ESBUILD")
 esbuild.build({
     entryPoints: ["src/index.ts"],
